+++ previous-versions/banking/src/sdk/models/shared/errormessage.ts
@@ -6,26 +6,44 @@ import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { Expose } from "class-transformer";
 
 export class ErrorMessage extends SpeakeasyBase {
+    /**
+     * `True` if the error occurred transiently and can be retried.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "canBeRetried" })
     canBeRetried?: string;
 
+    /**
+     * Unique identifier used to propagate to all downstream services and determine the source of the error.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "correlationId" })
     correlationId?: string;
 
+    /**
+     * Machine readable error code used to automate processes based on the code returned.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "detailedErrorCode" })
     detailedErrorCode?: number;
 
+    /**
+     * A brief description of the error.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "error" })
     error?: string;
 
+    /**
+     * Codat's service the returned the error.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "service" })
     service?: string;
 
+    /**
+     * The HTTP status code returned by the error.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "statusCode" })
     statusCode?: number;