+++ previous-versions/banking/src/sdk/models/shared/index.ts
@@ -4,11 +4,7 @@
 
 export * from "./account";
 export * from "./accountbalance";
-export * from "./accountbalanceamounts";
 export * from "./accountbalances";
-export * from "./accountidentifiers";
-export * from "./accountidentifiertype";
-export * from "./accountinstitution";
 export * from "./accounts";
 export * from "./accounttype";
 export * from "./errormessage";