+++ previous-versions/banking/RELEASES.md
@@ -539,3 +539,13 @@ Based on:
 - [typescript v0.29.0] previous-versions/banking
 ### Releases
 - [NPM v0.29.0] https://www.npmjs.com/package/@codat/banking/v/0.29.0 - previous-versions/banking
+
+## 2023-10-09 01:22:49
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Banking.yaml
+- Speakeasy CLI 1.96.1 (2.150.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.30.0] previous-versions/banking
+### Releases
+- [NPM v0.30.0] https://www.npmjs.com/package/@codat/banking/v/0.30.0 - previous-versions/banking